Dynamic mode now handles non-existing files
The current status is the following:

Given a page uses a document’s ID as parameter
And an other page which perform a search and create
  links to the first one using fetched IDs,
If a document exists only in a specific ref,
Then not page will be generated for its ID.

The problem is that it’s possible to provide
access_token/ref and see this particular
document… but we need to start from an other place
(in order to find a valid file which will load the
lib).

This commit create a fallback: if the page is not
found, the HTTP server only has to serve an other
simple static HTML file, which loads the lib.
At this point, the lib will find the right page to
load, fetch its template and render it.

WARNING: In case of route’s overloading (if a path
could be handled by 2 routes), there is currently
no way to specify the priority.