Stop using jsdom
The main library now use regular expressions instead
of using DOM helpers, so it’s not mandatory anymore
to provide a DOM implementation.

This allows to use characters “<” and “&” in templates
(which were escaped by jsom).

Notes: it is now possible to change template markers
by “<%” and “%>”, we may introduce new syntax later.